requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
beautifulsoup4==4.7.0
billiard==3.5.0.4
boto==2.49.0
boto3==1.9.49
botocore==1.12.50
bs4==0.0.1
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
docutils==0.14
future==0.16.0
idna==2.7
jmespath==0.9.3
kombu==4.2.1
pycparser==2.19
pymongo==3.7.2
python-dateutil==2.7.5
python-telegram-bot==11.1.0
pytz==2018.7
redis==2.10.5
requests==2.21.0
s3transfer==0.1.13
six==1.11.0
soupsieve==1.6.2
urllib3==1.24.1
vine==1.1.4
